+++ src/grepros/outputs.py
@@ -356,7 +356,6 @@ class BagSink(SinkBase):
         @param   args.META      whether to print metainfo
         @param   args.OUTFILE   name of ROS bagfile to create or append to
         @param   args.VERBOSE   whether to print debug information
-        @param   args.PROGRESS  whether progress bar is being printed
         """
         super(BagSink, self).__init__(args)
         self._bag = None
@@ -366,7 +365,6 @@ class BagSink(SinkBase):
 
     def emit(self, topic, index, stamp, msg, match):
         """Writes message to output bagfile."""
-        meta_printed = not self._bag and self._args.VERBOSE
         if not self._bag:
             if self._args.VERBOSE:
                 sz = os.path.isfile(self._args.OUTFILE) and os.path.getsize(self._args.OUTFILE)
@@ -375,8 +373,7 @@ class BagSink(SinkBase):
             self._bag = rosapi.create_bag_writer(self._args.OUTFILE)
 
         if topic not in self._counts and self._args.VERBOSE:
-            inter = "\n" if not meta_printed and self._args.PROGRESS else ""
-            ConsolePrinter.debug("%sAdding topic %s.", inter, topic)
+            ConsolePrinter.debug("Adding topic %s.", topic)
 
         self._bag.write(topic, msg, stamp)
         super(BagSink, self).emit(topic, index, stamp, msg, match)
@@ -407,7 +404,6 @@ class CsvSink(SinkBase):
                                 will add topic name like "name.__my__topic.csv" for "/my/topic",
                                 will add counter like "name.__my__topic.2.csv" if exists
         @param   args.VERBOSE   whether to print debug information
-        @param   args.PROGRESS  whether progress bar is being printed
         """
         super(CsvSink, self).__init__(args)
         self._filebase      = args.OUTFILE  # Filename base, will be made unique
@@ -462,8 +458,7 @@ class CsvSink(SinkBase):
             w = csv.writer(f)
             if topic not in self._files:
                 if self._args.VERBOSE:
-                    inter = "\n" if self._args.PROGRESS else ""
-                    ConsolePrinter.debug("%sCreating %s.", inter, name)
+                    ConsolePrinter.debug("Creating %s.", name)
                 header = [topic + "/" + ".".join(map(str, p)) for p, _ in self._iter_fields(msg)]
                 metaheader = ["__time", "__datetime", "__type"]
                 w.writerow(self._format_row(metaheader + header))
@@ -525,7 +520,6 @@ class HtmlSink(SinkBase, TextSinkMixin):
                                          both sides if one value, start and end if more than one,
                                          or no wrapping if zero values
         @param   args.ORDERBY            "topic" or "type" if any to group results by
-        @param   args.PROGRESS           whether progress bar is being printed
         """
         args = copy.deepcopy(args)
         args.WRAP_WIDTH = self.WRAP_WIDTH
@@ -616,8 +610,7 @@ class HtmlSink(SinkBase, TextSinkMixin):
                 break  # while
             (topic, index, stamp, msg, match) = entry
             if self._args.VERBOSE and topic not in self._counts:
-                inter = "\n" if self._args.PROGRESS else ""
-                ConsolePrinter.debug("%sAdding topic %s.", inter, topic)
+                ConsolePrinter.debug("Adding topic %s.", topic)
             yield entry
             super(HtmlSink, self).emit(topic, index, stamp, msg, match)
         try:
@@ -734,7 +727,6 @@ class SqliteSink(SinkBase, TextSinkMixin):
                                    will be appended to if exists
         @param   args.WRAP_WIDTH   character width to wrap message YAML output at
         @param   args.VERBOSE      whether to print debug information
-        @param   args.PROGRESS     whether progress bar is being printed
         """
         args = TextSinkMixin.make_full_yaml_args(args)
 
@@ -825,8 +817,7 @@ class SqliteSink(SinkBase, TextSinkMixin):
             fargs["name"] = quote(topic + suffix)
             sql = self.CREATE_TOPIC_VIEW % fargs
             if self._args.VERBOSE:
-                inter = "\n" if self._args.PROGRESS else ""
-                ConsolePrinter.debug("%sAdding topic %s.", inter, topic)
+                ConsolePrinter.debug("Adding topic %s.", topic)
             self._db.execute(sql)
             self._schema["view"].setdefault(topic, {})[typename] = True
 
@@ -946,7 +937,6 @@ class TopicSink(SinkBase):
         @param   args.PUBLISH_FIXNAME   single output topic name to publish to,
                                         overrides prefix and suffix if given
         @param   args.VERBOSE           whether to print debug information
-        @param   args.PROGRESS          whether progress bar is being printed
         """
         super(TopicSink, self).__init__(args)
         self._pubs = {}  # {(intopic, cls): ROS publisher}
@@ -959,8 +949,7 @@ class TopicSink(SinkBase):
             topic2 = self._args.PUBLISH_PREFIX + topic + self._args.PUBLISH_SUFFIX
             topic2 = self._args.PUBLISH_FIXNAME or topic2
             if self._args.VERBOSE:
-                inter = "\n" if self._args.PROGRESS else ""
-                ConsolePrinter.debug("%sPublishing from %s to %s.", inter, topic, topic2)
+                ConsolePrinter.debug("Publishing from %s to %s.", topic, topic2)
 
             pub = None
             if self._args.PUBLISH_FIXNAME: