+++ CHANGELOG.rst
@@ -4,7 +4,7 @@ Changelog for package grepros
 
 0.2.5 (2021-11-27)
 -------------------
-* add --write-progress option
+* add --progress option
 * match anything by default if no patterns given
 * add timeline to HTML output
 * auto-detect output format from given filename